Fix API_PATH bug and refactor things
Closes #20
